Big incoming changes to data structure. CSV seeder in JS made for testing purposes.
The data that will be displayed will contain different fields and it will be measured differently in the graphs. The adaptation to the data's new standards will take some time and I will probably have to write a bunch of stuff from scratch, but I am very satisfied with the progress I've made with this seeder.